+++ src/main/java/database/Page.java
@@ -1,4 +1,4 @@
-package database.page;
+package database;
 
 import java.io.Serializable;
 import java.util.ArrayList;
